add interface and more files
