+++ app/src/main/java/com/library/library_objects.kt
@@ -1,0 +1,17 @@
+package com.library
+
+abstract class LibraryObjects(
+    val objectId: Int,
+    var access: Boolean,
+    val name: String
+): Returnable{
+    fun smallInformation(){
+        if (access){
+            println("$name доступна: Да")
+        } else {
+            println("$name доступна: Нет")
+        }
+    }
+
+    abstract fun longInformation()
+}
