+++ app/src/main/java/com/library/logic_for_get_home.kt
@@ -1,0 +1,5 @@
+package com.library
+
+interface Home_able {
+    fun getHome()
+}
