+++ app/src/main/java/com/library/logic_for_read.kt
@@ -1,0 +1,5 @@
+package com.library
+
+interface Readable{
+    fun getRead()
+}
