+++ app/src/main/java/com/library/logic_for_return.kt
@@ -1,0 +1,5 @@
+package com.library
+
+interface Returnable {
+    fun returnObject()
+}
